# Start optimising the AOT module.