… tests/c/funcptrarg_noir.c.O0
… tests/c/nested_writetoptr.c
… tests/c/promote_not_great.c