**Companion PR: https://github.com/ykjit/ykllvm/pull/74**

Until now we've required the interpreter to be built with zero AOT optimisation. This is because we can't tolerate optimisations that change the block structure in the codegen backend: this would cause our mapper to break.

That's not to say we can't do *any* optimisation.

This change turns on one AOT optimisation: the instrcombine pass: https://llvm.org/doxygen/InstructionCombining_8cpp_source.html

(It does some simplifications and constant folding etc)

There are some fiddly details:

 - We still can't let the backend optimise things. A new flag, `-yk-optnone-after-ir-passes`, was added to prevent it from doing so. Even at -O0 the backend was optimising stuff.

 - The user is still required to build with O0. We add our pass on top of "no optimisations". Using higher optimisation levels right now will cause chaos.

 - This shook out a couple fo bugs in the trace compiler -- also fixed in this change.

I hope that in the future we can run with a full -O2 pipeline, just with backend optimisation turned off, but currently the trace compiler makes too many assumptions about the shape of the IR, and optimisations break those assumptions.

This way we can turn on individual optimisations and work our way up incrementally.