@vext01 2023-07-11
Question: should we run tests with and without our curated set of optimisations enabled? Currently the optimisations are hard-coded on in `yk-config`.