@ltratt 2023-07-11
> Question: should we run tests with and without our curated set of optimisations enabled?

Good idea -- ideally, yes! But we can do that in the next PR, rather than weigh this one down (unless it's easy to do).