@vext01 2023-07-12
@ptersilie [103b3e6](https://github.com/ykjit/yk/pull/779/commits/103b3e6766ab4a6cf72b697b2017fca86bf830ad) adds `ctpop` to the whitelist, and i've reverted the llvm change from https://github.com/ykjit/ykllvm/pull/74.

OK to squash (both)?