@vext01 2023-07-12
FWIW, this didn't make a great deal of difference to performance according to ykbench:

# master branch

```
Benchmark 1: /home/vext01/research/yklua/src/lua ./yklua/fannkuchredux.lua 3
  Time (mean ± σ):      2.597 s ±  0.044 s    [User: 2.502 s, System: 0.066 s]
  Range (min … max):    2.532 s …  2.650 s    10 runs
 
Benchmark 1: /home/vext01/research/yklua/src/lua ./yklua/fasta.lua 10
  Time (mean ± σ):      2.874 s ±  0.034 s    [User: 2.763 s, System: 0.059 s]
  Range (min … max):    2.824 s …  2.948 s    10 runs
```
 
# this branch

```
Benchmark 1: /home/vext01/research/yklua/src/lua ./yklua/fannkuchredux.lua 3
  Time (mean ± σ):      2.563 s ±  0.034 s    [User: 2.465 s, System: 0.068 s]
  Range (min … max):    2.511 s …  2.632 s    10 runs
 
Benchmark 1: /home/vext01/research/yklua/src/lua ./yklua/fasta.lua 10
  Time (mean ± σ):      2.831 s ±  0.049 s    [User: 2.734 s, System: 0.057 s]
  Range (min … max):    2.776 s …  2.950 s    10 runs
```


This was using a release build of yk, but with the `yk_testing` feature enabled so we can do `YKD_SERIALISE_COMPILATION`.

Took me a while to realise that `YKD_SERIALISE_COMPILATION` is guarded like that. Maybe not desirable?