@vext01 2023-07-12
splat.

(Left the second commit separate, since it's not needed for our test suite, only for yklua)